Updated Qsys DSP plugin - Updated dependencies to support .net8. Adjusted control classes and base QsyDspTcp class to meet the new API of the QscQsys dependency.

+++ pkd-audioDsp-plugins/QscDsp/QscZoneEnable.cs
@@ -75,7 +75,7 @@
 			if (_zoneToggles.TryGetValue(zoneId, out var target))
 			{
 
-				int newState = target.CurrentState ? 1 : 0;
+				ushort newState = (ushort)(target.CurrentState ? 1 : 0);
 				target.SetBoolean(newState);
 			}
 		}
